Added logic to handleGetHelpRequest to check for game in progress
Added logic to handleGetHelpRequest in ReindeerGames and ChemistryFlashCards samples to fix a bug that occurs when triggering the Help Intent when invoking the skills. Logic added checks if a game is in progress, if no game is in progress it will prompt user to start a game, else will provide help.